+++ src/commands/hardis/project/deploy/sources/dx.ts
@@ -362,7 +362,7 @@ If you need to increase the deployment waiting time (force:source:deploy --wait 
     }
 
     // Process deployment (or deployment check)
-    const { messages } = await forceSourceDeploy(packageXmlFile, this.checkOnly, testlevel, this.debugMode, this, forceSourceDeployOptions);
+    const { messages, quickDeploy } = await forceSourceDeploy(packageXmlFile, this.checkOnly, testlevel, this.debugMode, this, forceSourceDeployOptions);
 
     // Set ListViews to scope Mine if defined in .sfdx-hardis.yml
     if (this.configInfo.listViewsToSetToMine && this.checkOnly === false) {
@@ -379,7 +379,7 @@ If you need to increase the deployment waiting time (force:source:deploy --wait 
         branchMd = UtilsNotifs.markdownLink(branchUrl, currentGitBranch);
       }
       let notifMessage = `Deployment has been successfully processed from branch ${branchMd} to org ${linkMarkdown}`;
-      notifMessage += delta ? " (🌙 delta deployment)" : " (🌕 full deployment)";
+      notifMessage += quickDeploy ? " (🚀 quick deployment)" : delta ? " (🌙 delta deployment)" : " (🌕 full deployment)";
       const notifButtons = [];
       const jobUrl = await GitProvider.getJobUrl();
       if (jobUrl) {